Added process initializer and initial process display

+++ src/process.cpp
@@ -15,7 +15,7 @@ int Process::Pid() { return pid_; }
 
 // TODO: Return this process's CPU utilization
 float Process::CpuUtilization() {
-    return LinuxParser::ActiveJiffies(pid_)/LinuxParser::Jiffies();
+    return (float)LinuxParser::ActiveJiffies(pid_)/(float)LinuxParser::Jiffies();
 }
 
 string Process::Command() { return LinuxParser::Command(pid_); }
